Move board HAL
